АпДеЙт
Убраны статик, классы раскиданы по файлам

Чётко Ясно Классно!
