+++ SnakeGame/SnakeGame/GameState.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SnakeGame
+{
+    enum GameState
+    {
+        MainMenu,
+        InGame,
+        GameOver,
+        Paused
+    }
+}
