+++ SnakeGame/SnakeGame/bin/Debug/gameState.json
@@ -1,1 +1,1 @@
-{"Snake":null,"Food":{"X":0,"Y":0,"Color":0,"PixelSize":0},"Score":0,"PlayerName":null}
+{"IsSavedGame":false,"Food":{"X":0,"Y":0,"Color":0,"PixelSize":0},"Score":0,"PlayerName":null,"Head":{"X":0,"Y":0,"Color":0,"PixelSize":0},"Body":null}
