+++ SnakeGame/SnakeGame/bin/Debug/records.txt
@@ -4,7 +4,7 @@ storkych 25
  10
  10
  9
- 7
- 7
-. 7
- 6
+ddd 8
+ddd 8
+ddddddd 8
+ddddddd 8
